:sparkles: 12 column grid container and classes
This component and it's related css classes should create the foundation of a fluidly responsive 960px wide 12 column grid system. As the grid compresses the columns and gutters should decrease in width while maintaing their relative proportions. Grids should also be able to be nested inside of each other.

Addresses #94